requirements.txt
autopep8==2.0.0
numpy==1.23.5
Pillow==9.3.0
pycodestyle==2.10.0
tomli==2.0.1
